+++ packages/core/client/User.js
@@ -31,7 +31,6 @@ Kepler.User = Class.extend({
 	templatePanel: 'panelPlace',	
 	templatePopup: 'popupUser',
 	templateMarker: 'markerUser',
-	iconClassName: 'marker-profile',
 	data: {},
 	
 	init: function(userId) {
@@ -100,8 +99,7 @@ Kepler.User = Class.extend({
 				/*conSize: new L.Point(iconOpts.iconSize),
 				iconAnchor: new L.Point(iconOpts.iconAnchor),
 				popupAnchor: new L.Point(iconOpts.popupAnchor),*/
-				nodeHtml: L.DomUtil.create('div'),
-				className: self.iconClassName
+				nodeHtml: L.DomUtil.create('div')
 			});
 			self.marker = new L.Marker(self.loc, {icon: self.icon});
 			self.marker.item = self;