+++ CHANGELOG.md
@@ -3,6 +3,7 @@ To see all releases [here](https://github.com/Keplerjs/Kepler/releases)
 
 ## Next version
 - support templates custom ordering
+- moved Status.js from core to lib
 
 ## v1.2.5
 - plugins renamed dirs in plain name